Validaciones de Eliminar Producto

+++ src/app/components/productos/view-products/view-products.component.ts
@@ -37,7 +37,9 @@ export class ViewProductsComponent implements OnInit{
  
 
   deleteProduct(key){
-    this.productService.deleteProduct(key);
+    if(confirm('¿Seguro que desea eliminar este producto?')){
+      this.productService.deleteProduct(key);
+    }
   }
